requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
astroid==2.12.13
async-timeout==4.0.2
attrs==22.1.0
charset-normalizer==2.1.1
dill==0.3.6
frozenlist==1.3.3
idna==3.4
isort==5.10.1
lazy-object-proxy==1.8.0
mccabe==0.7.0
multidict==6.0.2
platformdirs==2.5.4
pylint==2.15.7
python-dotenv==0.21.0
pytz==2022.6
slackclient==2.9.4
tomli==2.0.1
tomlkit==0.11.6
tqdm==4.64.1
typing_extensions==4.4.0
wrapt==1.14.1
yarl==1.8.1
